Revert "Set order to float"
This reverts commit 673475d1cc9b6e1c536978e597e4a007c5388387.

+++ lib/FilterParametersWidget.py
@@ -7,7 +7,7 @@ from PyQt6.QtWidgets import (
 
 class FilterParametersWidget(QGroupBox):
     def __init__(self,
-        order: float,
+        order: int,
         cutoff: float,
         *args, **kwargs
     ) -> None:
